requirements.txt
flask>=2.0,<3.0
flask-cors>=3.0
torch>=1.13.0,<2.0
torchvision>=0.14.0,<1.0
pillow>=8.0
numpy>=1.21
gunicorn>=20.0
python-dotenv>=0.19